Move the email preview button component to its own file
